Move verifyJwtFromHeader() into its own singleton class
This will make it easier to mock.

Addresses #17, #20.
